move vault balance chart
