+++ packages/webapp/src/routes/vaults/[network]-[id]/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { VaultBalanceChangesTable } from '@rainlanguage/ui-components';
+	import { VaultBalanceChangesTable, VaultBalanceChart } from '@rainlanguage/ui-components';
 	import { page } from '$app/stores';
 
 	const { settings } = $page.data.stores;
@@ -8,4 +8,5 @@
 	const subgraphUrl = $settings?.subgraphs?.[network] || '';
 </script>
 
+<VaultBalanceChart {id} {subgraphUrl} />
 <VaultBalanceChangesTable {id} {subgraphUrl} />
